Add files via upload
Project: Power BI Survey Dashboard.
The primary goal of this project is to develop an interactive dashboard using Power BI that enables users to input their responses to survey questions related to happiness and salaries. The dashboard will then visualize this data in an intuitive and insightful manner, allowing stakeholders to derive meaningful conclusions and make informed decisions.